+++ lib/cli/src/generators/WEB-COMPONENTS/index.ts
@@ -2,9 +2,7 @@ import { baseGenerator, Generator } from '../baseGenerator';
 import { copyTemplate } from '../../helpers';
 
 const generator: Generator = async (packageManager, npmOptions, options) => {
-  baseGenerator(packageManager, npmOptions, options, 'web-components', {
-    extraPackages: ['lit-html'],
-  });
+  await baseGenerator(packageManager, npmOptions, options, 'web-components');
   copyTemplate(__dirname, options.storyFormat);
 };