+++ lib/components/src/controls/options/Checkbox.tsx
@@ -3,6 +3,7 @@ import { styled } from '@storybook/theming';
 import { logger } from '@storybook/client-logger';
 import { ControlProps, OptionsMultiSelection, NormalizedOptionsConfig } from '../types';
 import { selectedKeys, selectedValues } from './helpers';
+import { getControlId } from '../helpers';
 
 const Wrapper = styled.div<{ isInline: boolean }>(({ isInline }) =>
   isInline
@@ -69,10 +70,12 @@ export const CheckboxControl: FC<CheckboxProps> = ({
     setSelected(updated);
   };
 
+  const controlId = getControlId(name);
+
   return (
     <Wrapper isInline={isInline}>
-      {Object.keys(options).map((key) => {
-        const id = `${name}-${key}`;
+      {Object.keys(options).map((key, index) => {
+        const id = `${controlId}-${index}`;
         return (
           <Label key={id} htmlFor={id}>
             <input