+++ lib/core-server/src/index.ts
@@ -16,5 +16,3 @@ export {
 
 export * from './build-static';
 export * from './build-dev';
-
-export const managerPreset = require.resolve('./presets/manager-preset');